python-backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
python-magic
python-dotenv
numpy>=1.24.0
scipy>=1.10.0
soundfile
librosa
pyloudnorm
gunicorn
supabase
python-magic-bin; sys_platform == 'win32'
demucs
spleeter
pydub
torch
torchvision
torchaudio
b2sdk
